requirements.txt
# 本地语音助手依赖清单
# Local Voice Assistant Requirements

# === 核心依赖 ===
# Core dependencies
pyaudio>=0.2.13
pygame>=2.5.0
edge-tts>=6.1.9
numpy>=1.24.0

# === ASR 语音识别 ===
# ASR (Automatic Speech Recognition)
funasr[onnxruntime]>=1.0.0
soundfile>=0.12.0
sounddevice>=0.4.6

# === LLM 大语言模型 ===
# LLM (Large Language Model)
llama-cpp-python>=0.2.0
torch>=2.0.0
transformers>=4.35.0

# === VAD 语音活动检测 ===
# VAD (Voice Activity Detection)
webrtcvad>=2.0.10

# === 语言检测 ===
# Language Detection
langid>=1.1.6
langdetect>=1.0.9

# === 系统依赖 (需用 apt 安装) ===
# System dependencies (install with apt)
# sudo apt-get install portaudio19-dev python3-pyaudio
